Minimal repro for TransportError.
This code is minimally extracted from another code base and is capable
of exhibiting an issue where Quinn encounters FLOW_CONTROL_ERROR on the
client side.

It is unclear to me currently why this is happening, nor is it clear why
it happens only some of the time. It seems to have something to do with
multiple futures all trying to read from the filesystem, although all of
those could be red herrings.